services/trade_to_ohlc/pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["deepsuchak <suchakdeep@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
quixstreams = "^2.5.1"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.3.3"
loguru = "^0.7.2"
ruff = "0.4.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 88

[tool.ruff.format]
quote-style = "single"
indent-style = "space"
docstring-code-format = true

[tool.ruff.lint]
extend-select = ["I"]
